+++ packages-exp/installations-types-exp/index.d.ts
@@ -16,6 +16,8 @@
  */
 
 /**
+ * Public interface of the FirebaseInstallations SDK.
+ *
  * @public
  */
 export interface FirebaseInstallations {}