+++ packages/firestore/lite/src/api/reference.ts
@@ -109,10 +109,10 @@ export class DocumentReference<T = DocumentData> extends _DocumentKeyReference<
   /** @hideconstructor */
   constructor(
     readonly firestore: FirebaseFirestore,
-    readonly converter: FirestoreDataConverter<T> | null,
+    _converter: FirestoreDataConverter<T> | null,
     readonly _path: ResourcePath
   ) {
-    super(firestore._databaseId, new DocumentKey(_path), converter);
+    super(firestore._databaseId, new DocumentKey(_path), _converter);
   }
 
   get id(): string {
@@ -157,7 +157,7 @@ export class Query<T = DocumentData> {
   /** @hideconstructor protected */
   constructor(
     readonly firestore: FirebaseFirestore,
-    readonly converter: FirestoreDataConverter<T> | null,
+    readonly _converter: FirestoreDataConverter<T> | null,
     readonly _query: InternalQuery
   ) {}
 
@@ -220,7 +220,7 @@ class QueryFilterConstraint extends QueryConstraint {
     );
     return new Query(
       query.firestore,
-      query.converter,
+      query._converter,
       queryWithAddedFilter(query._query, filter)
     );
   }
@@ -230,11 +230,13 @@ export type WhereFilterOp =
   | '<'
   | '<='
   | '=='
+  | '!='
   | '>='
   | '>'
   | 'array-contains'
   | 'in'
-  | 'array-contains-any';
+  | 'array-contains-any'
+  | 'not-in';
 
 export function where(
   fieldPath: string | FieldPath,
@@ -262,7 +264,7 @@ class QueryOrderByConstraint extends QueryConstraint {
     const orderBy = newQueryOrderBy(query._query, this._field, this._direction);
     return new Query(
       query.firestore,
-      query.converter,
+      query._converter,
       queryWithAddedOrderBy(query._query, orderBy)
     );
   }
@@ -293,7 +295,7 @@ class QueryLimitConstraint extends QueryConstraint {
   _apply<T>(query: Query<T>): Query<T> {
     return new Query(
       query.firestore,
-      query.converter,
+      query._converter,
       queryWithLimit(query._query, this._limit, this._limitType)
     );
   }
@@ -327,7 +329,7 @@ class QueryStartAtConstraint extends QueryConstraint {
     );
     return new Query(
       query.firestore,
-      query.converter,
+      query._converter,
       queryWithStartAt(query._query, bound)
     );
   }
@@ -373,7 +375,7 @@ class QueryEndAtConstraint extends QueryConstraint {
     );
     return new Query(
       query.firestore,
-      query.converter,
+      query._converter,
       queryWithEndAt(query._query, bound)
     );
   }
@@ -466,7 +468,7 @@ export class CollectionReference<T = DocumentData> extends Query<T> {
     validateNonEmptyArgument('CollectionReference.doc', 'path', path);
     const absolutePath = this._path.child(ResourcePath.fromString(path!));
     validateDocumentPath(absolutePath);
-    return new DocumentReference(this.firestore, this.converter, absolutePath);
+    return new DocumentReference(this.firestore, this._converter, absolutePath);
   }
 
   withConverter<U>(
@@ -592,7 +594,7 @@ export function doc<T>(
     validateDocumentPath(absolutePath);
     return new DocumentReference(
       parent.firestore,
-      parent instanceof CollectionReference ? parent.converter : null,
+      parent instanceof CollectionReference ? parent._converter : null,
       absolutePath
     );
   }
@@ -627,7 +629,7 @@ export function getDocs<T>(query: Query<T>): Promise<QuerySnapshot<T>> {
           query.firestore,
           doc.key,
           doc,
-          query.converter
+          query._converter
         )
     );
 
@@ -738,7 +740,10 @@ export function addDoc<T>(
 ): Promise<DocumentReference<T>> {
   const docRef = doc(reference);
 
-  const convertedValue = applyFirestoreDataConverter(reference.converter, data);
+  const convertedValue = applyFirestoreDataConverter(
+    reference._converter,
+    data
+  );
 
   const dataReader = newUserDataReader(reference.firestore);
   const parsed = parseSetData(
@@ -769,7 +774,7 @@ export function refEqual<T>(
     return (
       left.firestore === right.firestore &&
       left.path === right.path &&
-      left.converter === right.converter
+      left._converter === right._converter
     );
   }
   return false;
@@ -780,7 +785,7 @@ export function queryEqual<T>(left: Query<T>, right: Query<T>): boolean {
     return (
       left.firestore === right.firestore &&
       queryEquals(left._query, right._query) &&
-      left.converter === right.converter
+      left._converter === right._converter
     );
   }
   return false;